ai_discussion/requirements.txt
# ai_discussion - AI 토론 (LangGraph + RAG)
streamlit>=1.28.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
wikipedia>=1.4.0
openai>=1.0.0
